Make expect_messages() work for more than one message emitted
Fixes #51. 

The `tryCatch()` code does not permit to check all the messages that are emitted by the expression in `current`. By replacing it with `withCallingHandlers()`, it is possible to accumulate all messages emitted by the evaluated expression (here, into `testmsg`). In a subsequent step, the concatenated string of messages in `testmsg` is checked as usual and the return values are set.